requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
sqlalchemy
chromadb
openai
beautifulsoup4
psycopg2-binary
pydantic-settings
